refactor: ユーザー関連のatomをusersディレクトリに移動
