+++ src/features/users/components/SearchInput.tsx
@@ -1,8 +1,9 @@
 // src/features/components/SearchInput.tsx
 import { TextInput } from "@mantine/core";
 import { IconSearch } from "@tabler/icons-react";
-import { filterInputAtom } from "atoms/user/userInfoAtom";
 import { useAtom } from "jotai";
+
+import { filterInputAtom } from "../atoms/userInfoAtom";
 
 const SearchInput = () => {
   const [filterInput, setFilterInput] = useAtom(filterInputAtom);
